counter button 추가. buttons 파일 제거. 접근제어자 지정
